Code clean-up across the entire project.  Includes the following non-functional changes: * scala convention fixes - method parens, method bodies for unit functions, type shadowing, extraneous semicolons, etc. * spelling mistakes - a few variable name fixes, but mostly in comments * scaladoc fixes - missing param descriptions, malformed html, etc. * copyright fixes - looks like a search and replace munged a bunch of the license headers, so I fixed this
Also did 1 functional change?  the control test wasn't working due to an NPE in an implicit conversion, so I added a null check.  This sort of check may be needed in other places (like everywhere) to let you remove properties by setting to null, but I didn't want to tackle that in this check-in.